Add gulp eslint task
This change is for issue #134

This change adds the gulp task for eslint. Currently it is just
using the default `eslint()` configuration, but the configuration
is customizable in the gulp task as well as adding .eslintrc files
at any level of the directory structure.

Curently there are 12 errors after running eslint. Not sure why these
are errors as the syntax is correct.

Once this test is working completely it can be added to the default
task set or continuous integration task set.

Usage:
  $ gulp eslint

Documentation:
  https://github.com/adametry/gulp-eslint
  http://eslint.org/docs/user-guide/configuring